Add currency stories to the readme
First-class currency support is going to be included in the rewrite,
as it's a challenging feature to add unless it's there from the
beginning. Furthermore, there are quite a lot of non-USD loans
on r/borrow and the current approach leads to a reasonable amount
of confusion for new lenders.